// cypress/interfaces/user-data.ts
export interface UserData {
  name: string;
  email: string;
  password: string;
  title: string;
  birthDate: string;
  birthMonth: string;
  birthYear: string;
  firstName: string;
  lastName: string;
  company: string;
  address: string;
  address2: string;
  country: string;
  zipcode: string;
  state: string;
  city: string;
  mobile: string;
}